Struct josephine::JSContext

pub struct JSContext<S> { /* fields omitted */ }

The type for JS contexts whose current state is S.

Methods

impl JSContext<Owned>

pub fn new() -> JSContext<Owned>

Create a new JSContext.

impl<S> JSContext<S>

pub fn snapshot<'a>(&'a mut self) -> JSContext<Snapshotted<'a, S>>

Get a snapshot of the JS state. The snapshot only allows access to the methods that are guaranteed not to call GC, so we don't need to root JS-managed pointers during the lifetime of a snapshot.

pub fn enter_known_compartment<'a, 'b, C, T>(
&'a mut self,
managed: JSManaged<'b, C, T>
) -> JSContext<Entered<'a, C, T::Aged, S>> where
T: JSRootable<'a>,
C: Compartment,

Enter a known compartment.

pub fn enter_unknown_compartment<'a, 'b, C, T>(
&'a mut self,
managed: JSManaged<'b, C, T>
) -> JSContext<Entered<'a, BOUND<'a>, T::Aged, S>> where
T: JSRootable<'a>,

Enter a compartment.

pub fn manage<'a, C, T>(&'a mut self, value: T) -> JSManaged<'a, C, T::Aged> where
S: CanAlloc + InCompartment<C>,
C: Compartment,
T: JSTraceable + JSInitializable + JSRootable<'a>,

Give ownership of data to JS. This allocates JS heap, which may trigger GC.

pub fn snapshot_manage<'a, C, T>(
&'a mut self,
value: T
) -> (JSContext<Snapshotted<'a, S>>, JSManaged<'a, C, T::Aged>) where
S: CanAlloc + InCompartment<C>,
C: Compartment,
T: JSTraceable + JSInitializable + JSRootable<'a>,

Give ownership of data to JS. This allocates JS heap, which may trigger GC.

pub fn create_compartment<'a, T>(
&'a mut self
) -> JSContext<Initializing<'a, BOUND<'a>, T>> where
S: CanCreateCompartments,
T: JSInitializable + JSTraceable,

Create a compartment

pub fn global_manage<'a, C, T>(
self,
value: T
) -> JSContext<Initialized<'a, C, T::Aged>> where
S: IsInitializing<'a, C, T>,
T: JSTraceable + JSRootable<'a> + JSTransplantable<C, Transplanted = T>,

Finish initializing a JS Context

pub fn create_global<'a, T>(
&'a mut self,
value: T
) -> JSContext<Initialized<'a, BOUND<'a>, T::Aged>> where
S: CanCreateCompartments,
T: JSInitializable + JSTraceable + JSRootable<'a> + JSTransplantable<BOUND<'a>, Transplanted = T>,

Shortcut to create a compartment and finish initializing in one go.

pub fn entered<'a, C, T>(&self) -> JSManaged<'a, C, T> where
S: IsEntered<'a, C, T>,

Get the object we entered the current compartment via

pub fn global<'a, C, T>(&self) -> JSManaged<'a, C, T> where
S: IsInitialized<'a, C, T>,

Get the global of an initialized context.

pub fn new_root<'a, 'b, T>(&'b mut self) -> JSRoot<'a, T>

Create a new root.

pub fn cx(&self) -> *mut JSContext

pub fn rt(&self) -> *mut JSRuntime

pub fn gc(&mut self) where
S: CanAlloc,

pub fn evaluate<C>(&mut self, code: &str) -> Result<JSVal, JSEvaluateErr> where
S: InCompartment<C>,

pub fn take_pending_exception(&mut self) -> Result<(), JSEvaluateErr>