Struct josephine::JSString

pub struct JSString<'a, C> { /* fields omitted */ }

The type of JS-managed strings in the same zone as compartment C, with lifetime a. Rust is much happier with flat string representations, so we flatten strings when they come into Rust.

Methods

impl<'a, C> JSString<'a, C>

pub fn to_jsstring(self) -> *mut JSString

pub fn to_jsflatstring(self) -> *mut JSFlatString

pub fn to_jsval(self) -> JSVal

pub fn len(self) -> usize

pub fn has_latin1_chars(self) -> bool

pub unsafe fn get_latin1_chars(self) -> &'a str

pub unsafe fn get_two_byte_chars(self) -> &'a [u16]

pub fn contents(self) -> JSStringContents<'a>

pub unsafe fn from_jsstring(js_string: *mut JSString) -> JSString<'a, C>

pub unsafe fn from_latin1_unchecked<S>(
cx: &'a mut JSContext<S>,
string: &str
) -> JSString<'a, C> where
S: CanAlloc + InCompartment<C>,

pub fn from_latin1<S>(
cx: &'a mut JSContext<S>,
string: &str
) -> Option<JSString<'a, C>> where
S: CanAlloc + InCompartment<C>,

pub fn from_twobyte<S>(
cx: &'a mut JSContext<S>,
slice: &[u16]
) -> JSString<'a, C> where
S: CanAlloc + InCompartment<C>,

pub fn from_str<S>(cx: &'a mut JSContext<S>, string: &str) -> JSString<'a, C> where
S: CanAlloc + InCompartment<C>,

pub fn clone_in<S, D>(self, cx: &'a mut JSContext<S>) -> JSString<'a, D> where
S: CanAlloc + InCompartment<D>,