Struct josephine::JSManaged

pub struct JSManaged<'a, C, T> { /* fields omitted */ }

The type of JS-managed data in a JS compartment C, with lifetime 'a and type T.

If the user has access to a JSManaged, then the JS-managed data is live for the given lifetime.

Methods

impl<'a, C, T> JSManaged<'a, C, T>

pub fn get<'b, S>(self, _: &'b JSContext<S>) -> T::Aged where
S: CanAccess,
C: Compartment,
T: JSRootable<'b>,
T::Aged: Copy,
'a: 'b,

Read-only access to JS-managed data.

pub fn borrow<'b, S>(self, _: &'b JSContext<S>) -> &'b T::Aged where
S: CanAccess,
C: Compartment,
T: JSRootable<'b>,
'a: 'b,

pub fn borrow_mut<'b, S>(self, _: &'b mut JSContext<S>) -> &'b mut T::Aged where
S: CanAccess,
C: Compartment,
T: JSRootable<'b>,
'a: 'b,

Read-write access to JS-managed data.

pub unsafe fn change_compartment<D>(self) -> JSManaged<'a, D, T::Transplanted> where
T: JSTransplantable<D>,

Change the compartment of JS-managed data.

pub unsafe fn change_lifetime<'b>(self) -> JSManaged<'b, C, T::Aged> where
T: JSRootable<'b>,

Change the lifetime of JS-managed data.

pub fn extend_lifetime<'b, 'c, S>(
self,
_: &'c JSContext<S>
) -> JSManaged<'b, C, T::Aged> where
S: CanExtend<'b>,
T: JSRootable<'b>,

It's safe to extend the lifetime of JS-managed data if it has been snapshotted.

pub fn forget_compartment(self) -> JSManaged<'a, SOMEWHERE, T::Transplanted> where
T: JSTransplantable<SOMEWHERE>,

Forget about which compartment the managed data is in. This is safe because when we mutate data in compartment C we require C: Compartment, which means it is never SOMEWHERE.

pub fn in_compartment<S, D>(
self,
cx: &JSContext<S>
) -> Option<JSManaged<'a, D, T::Transplanted>> where
T: JSTransplantable<D>,
S: InCompartment<D>,

Check to see if the current object is in the same compartment as another.

pub fn to_jsobject(self) -> *mut JSObject

pub fn to_jsval(self) -> JSVal